# Conflicts:
#	phonesin/src/main/resources/application.yml

+++ phonesin/src/main/java/com/regalaxy/phonesin/rental/model/repository/RentalRepositoryCustom.java
@@ -9,5 +9,5 @@ import java.util.List;
 public interface RentalRepositoryCustom {
     List<RentalDto> search(SearchDto searchDto);
 
-    RentalDetailDto detailInfo(int rental_id);
+    RentalDetailDto detailInfo(Long rental_id);
 }
